# openapi.yaml
openapi: 3.1.0
info:
  title: 企业短信管理平台 API
  version: 1.6.0
  description: |
    统一短信发送与管理平台对外接口。
    - /api/v1/messages/** 供内部系统调用，X-Api-Key 认证（双 Key 轮换宽限期内新旧均有效）
    - /api/v1/web/** 供 Vue3 前端调用，显式选择本地或 AD 认证源；15 分钟 access JWT 每次按数据库权威安全上下文校验，refresh JWT 七天内单次轮换
    - 所有时间均为 ISO8601 +08:00
    - v1.1：消息分类(category)、biz_id 幂等、失败重发、结果回调、双 Key 轮换、登出
    - v1.2：计费条(est_segments/quota_cost)、模板变量渲染(template_params)、号码级频控(removed_freq_limit)、
      测试发送(is_test)、营销同意(consent_confirmed)、批次列表/号码搜索/队列恢复/用户管理/回调密钥轮换
    - v1.3-v1.5：厂商{sN}模板转换、OTP打码、unmatched、异常检测、号码时间线、任务健康
    - v1.6：AUTH_MOCK、完整CRUD/计费预估/改期/授权解密、密文导出、owner/app角色与确定性Mock验收
    - v1.6.13：用户角色变更与强制下线使用数据库会话版本即时失效既有 JWT
    - v1.6.14：本地/AD Provider 显式路由、稳定主体 JWT 与首次强制改密
    - v1.6.24：统一 security_version 覆盖账号/身份/Provider/部门/角色，短 access JWT 与单次 refresh 轮换
    - v1.6.12：运行策略事务最终校验、动态过期时间与启动调度心跳一致
    - v1.6.11：报告回调引用新增无 PII 事件身份，raw 重放不重复投递同一报告
    - v1.6.28：callback/export 使用可续租 UUID fencing；回调暴露稳定 event_id 与租约状态
    - v1.6.33：拆分无依赖 livez 与严格依赖 readyz；健康响应不暴露内部组件或失败原因
    - v1.6.34：生产 callback 强制 HTTPS；签名密钥/协议版本按事件固化；敏感密文与导出帧绑定业务上下文
servers:
  - url: https://sms.example.com
...
paths:
...
  /api/v1/web/reports/dashboard:
    get:
      tags: [reports]
      summary: 仪表盘（今日分类别量/成功率/余额曲线/待审批/告警/uncertain数）
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/web/reports/stats:
    get:
      tags: [reports]
      summary: 统计报表（日/周/月 × 应用/部门 × 类别）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: granularity, in: query, required: false, schema: { type: string, enum: [day, week, month], default: day } }
        - { name: group_by, in: query, required: false, schema: { type: string, enum: [app, dept], default: app } }
        - { name: category, in: query, required: false, schema: { type: string, enum: [verify, notice, market, all], default: all } }
        - { name: start, in: query, required: false, schema: { anyOf: [{ type: string, format: date }, { type: 'null' }] } }
        - { name: end, in: query, required: false, schema: { anyOf: [{ type: string, format: date }, { type: 'null' }] } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportingModel' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/web/reports/export/{public_id}/download:
    get:
      tags: [reports]
      summary: 下载导出文件（UUID、稳定主体与部门授权；明文需单次二次认证）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: public_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Export-Step-Up, in: header, required: false, schema: { anyOf: [{ type: string, minLength: 1, maxLength: 256 }, { type: 'null' }] }, description: 仅明文导出必填；绑定账号、会话、IP与任务且单次使用 }
      responses:
        '200': { description: CSV 文件流 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/web/reports/export/{public_id}/step-up:
    post:
      tags: [reports]
      summary: 为指定明文导出签发五分钟单用途二次认证令牌
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: public_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportStepUpRequest' }
      responses:
        '200':
          description: 已签发
          content: { application/json: { schema: { $ref: '#/components/schemas/ExportStepUpResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '423': { description: 当前账号已锁定（ACCOUNT_LOCKED）, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/web/reports/export/{public_id}:
    get:
      tags: [reports]
      summary: 查询导出任务状态（稳定主体和显式角色/部门授权矩阵）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: public_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ExportTask' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/web/reports/export:
    post:
      tags: [reports]
      summary: 明细导出（异步，≤10万行；含明文手机号需权限并记审计）
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filters:
                  type: object
                  additionalProperties: false
                  properties:
                    start: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
                    end: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
                    category: { anyOf: [{ type: string, enum: [verify, notice, market] }, { type: 'null' }] }
                    status: { anyOf: [{ type: string, maxLength: 16 }, { type: 'null' }] }
                    app_id: { anyOf: [{ type: integer, minimum: 1 }, { type: 'null' }] }
                    dept: { anyOf: [{ type: string, maxLength: 128 }, { type: 'null' }] }
                    batch_no: { anyOf: [{ type: string, maxLength: 32 }, { type: 'null' }] }
                    phone: { anyOf: [{ type: string, pattern: '^1\d{10}$' }, { type: 'null' }], description: 请求后立即转phone_hmac，export_task.filters禁止保存明文 }
                decrypted: { type: boolean, default: false, description: 是否导出明文手机号（按角色，记审计） }
      responses:
        '202':
          description: 任务已创建
          content: { application/json: { schema: { $ref: '#/components/schemas/ExportTask' } } }

# ───────────────────────── Webhook（平台 → 应用，v1.1） ─────────────────────────
...
components:
  securitySchemes:
    MetricsBearerAuth:
      type: http
      scheme: bearer
      description: 仅供 Prometheus 抓取使用的独立、可轮换 secret。
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: 认证失败
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: 权限不足或审批回避
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: 资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    DashboardCategoryModel:
      type: object
      required: [category, total, total_segments, delivered, failed, unknown, success_rate]
      properties:
        category: { type: string, enum: [verify, notice, market] }
        total: { type: integer }
        total_segments: { type: integer }
        delivered: { type: integer }
        failed: { type: integer }
        unknown: { type: integer }
        success_rate: { type: number }
    DashboardBalancePointModel:
      type: object
      required: [stat_date, balance]
      properties:
        stat_date: { type: string, format: date }
        balance: { type: integer }
    DashboardAlertModel:
      type: object
      required: [level, title, created_at]
      properties:
        level: { type: string, enum: [info, warn, crit] }
        title: { type: string }
        created_at: { type: string, format: date-time }
    DashboardDispositionsModel:
      type: object
      required: [uncertain, unmatched, callback_dead]
      properties:
        uncertain: { type: integer }
        unmatched: { type: integer }
        callback_dead: { type: integer }
    DashboardChannelMonitorModel:
      type: object
      required: [realtime_queue, bulk_queue, qps_used, qps_rate, reserved_realtime_qps, stale]
      properties:
        realtime_queue: { anyOf: [{ type: integer }, { type: 'null' }] }
        bulk_queue: { anyOf: [{ type: integer }, { type: 'null' }] }
        qps_used: { anyOf: [{ type: integer }, { type: 'null' }] }
        qps_rate: { type: integer, minimum: 1 }
        reserved_realtime_qps: { type: integer, minimum: 0 }
        stale: { type: boolean }
    DashboardUiPolicyModel:
      type: object
      required: [test_send_max]
      properties:
        test_send_max: { type: integer, minimum: 1 }
    DashboardJobModel:
      type: object
      required: [job_name, last_run_at, last_status, stalled]
      properties:
        job_name: { type: string }
        last_run_at: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        last_status: { anyOf: [{ type: string, enum: [running, success, failed] }, { type: 'null' }] }
        stalled: { type: boolean }
    DashboardOperationsModel:
      type: object
      required: [current_balance, balances, alerts, dispositions, jobs, channel_monitor, balance_alert_threshold]
      properties:
        current_balance: { anyOf: [{ type: integer }, { type: 'null' }] }
        balances:
          type: array
          items: { $ref: '#/components/schemas/DashboardBalancePointModel' }
        alerts:
          type: array
          items: { $ref: '#/components/schemas/DashboardAlertModel' }
        dispositions: { $ref: '#/components/schemas/DashboardDispositionsModel' }
        jobs:
          type: array
          items: { $ref: '#/components/schemas/DashboardJobModel' }
        channel_monitor: { $ref: '#/components/schemas/DashboardChannelMonitorModel' }
        balance_alert_threshold: { type: integer, minimum: 1 }
    DashboardModel:
      type: object
      required: [refreshed_at, categories, overall_success_rate, pending_approvals, ui_policy]
      properties:
        refreshed_at: { type: string, format: date-time }
        categories:
          type: array
          items: { $ref: '#/components/schemas/DashboardCategoryModel' }
        overall_success_rate: { type: number }
        pending_approvals: { type: integer }
        ui_policy: { $ref: '#/components/schemas/DashboardUiPolicyModel' }
        operations:
          anyOf:
            - { $ref: '#/components/schemas/DashboardOperationsModel' }
            - { type: 'null' }
    ReportingRowModel:
      type: object
      required: [period_start, dim_value, dim_label, total, total_segments, delivered, failed, unknown, success_rate]
      properties:
        period_start: { type: string, format: date }
        dim_value: { type: string }
        dim_label: { type: string }
        total: { type: integer }
        total_segments: { type: integer }
        delivered: { type: integer }
        failed: { type: integer }
        unknown: { type: integer }
        success_rate: { type: number }
    ReportingModel:
      type: object
      required: [granularity, group_by, category, start, end, can_export_decrypted, summary, items]
      properties:
        granularity: { type: string, enum: [day, week, month] }
        group_by: { type: string, enum: [app, dept] }
        category: { type: string, enum: [verify, notice, market, all] }
        start: { type: string, format: date }
        end: { type: string, format: date }
        can_export_decrypted: { type: boolean }
        summary: { $ref: '#/components/schemas/ReportingSummaryModel' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ReportingRowModel' }
    ReportingSummaryModel:
      type: object
      required: [total, total_segments, delivered, failed, unknown, success_rate]
      properties:
        total: { type: integer }
        total_segments: { type: integer }
        delivered: { type: integer }
        failed: { type: integer }
        unknown: { type: integer }
        success_rate: { type: number }
...
    Error:
      type: object
      required: [code, message, detail]
      properties:
        code: { type: string, description: 见 CLAUDE.md 平台错误码表 }
        message: { type: string }
        detail: { type: object, nullable: true }
...
    ExportTask:
      type: object
      required: [id, status, decrypted, row_count, download_url, expires_at, created_at]
      properties:
        id: { type: string, format: uuid, description: 不可枚举公开任务ID；内部整数主键永不对外 }
        status: { type: string, enum: [pending, running, done, failed] }
        decrypted: { type: boolean }
        row_count: { anyOf: [{ type: integer }, { type: 'null' }] }
        download_url: { anyOf: [{ type: string }, { type: 'null' }], description: done且未过期时返回鉴权下载地址 }
        expires_at: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        created_at: { type: string, format: date-time }
    ExportStepUpRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password: { type: string, format: password, minLength: 1, maxLength: 1024, writeOnly: true }
    ExportStepUpResponse:
      type: object
      required: [token]
      properties:
        token: { type: string, description: 仅在浏览器易失内存中保存并单次使用 }
        expires_in: { type: integer, default: 300 }